add Rails 3 compatibility
