+++ init.rb
@@ -1,1 +1,1 @@
-require File.dirname(__FILE__) + '/rails/init'
+require File.join(File.dirname(__FILE__), "lib", "canonical_url")
